+++ src/backend/app/models/expense.py
@@ -1,7 +1,7 @@
 import datetime as dt
 
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, DateTime, Integer, ForeignKey
+from sqlalchemy import Column, DateTime, Integer, ForeignKey, Float
 
 from app.core.db import Base
 
@@ -11,4 +11,4 @@ class Expense(Base):
     date = Column(DateTime, default=dt.datetime.now, index=True)
     category_id = Column(Integer, ForeignKey('category.id'), nullable=False)
     category = relationship('Category', back_populates='expenses')
-    amount = Column(Integer, nullable=False)
+    amount = Column(Float, nullable=False)
